Add email validation and Separate the newsletter component
